Update groups/groups.class.php
bp_groups_setup_nav action is part of the default theme I suspect as it's not present on my install of 1.6.4 and we don't use the default theme.
changing it to bp_init did the trick, so I'm offering it up as a change. Without it, it would just 404.

Thanks!