Setting up mocks stubs for the Roles class
Previously the roles class was modified to avoid errors when BuddyPress isn't active, but a better approach is to leave as much of the core classes unmodified as possible and use the mock class instead. The previous commits that this reverts are e5a84f05080def255ec9cfec3cbe47f46a767d2a and 37d69c6b8897a96020d025ffbbadbfeee5637639.